idを項目として作成

+++ database/migrations/2020_11_07_044846_create_shop_coordinates_table.php
@@ -16,7 +16,7 @@ class CreateShopCoordinatesTable extends Migration
         Schema::connection('location')
             ->create('shop_coordinates', function (Blueprint $table) {
                 $table->id();
-                $table->increments('landmark_coordinate_id');
+                $table->bigInteger('landmark_coordinate_id');
                 $table->bigInteger('floor_coordinate_id');
                 $table->float('x1_coordinate');
                 $table->float('x2_coordinate');
